implement move command

+++ tests/src/main.rs
@@ -3,14 +3,16 @@ mod shuffle;
 
 use std::{
     net::SocketAddr,
-    path::{self, Path, PathBuf},
+    path::{Path, PathBuf},
     time::Duration,
 };
 
-use anyhow::{bail, Result};
+use anyhow::Result;
 use chrono::{DateTime, FixedOffset, Utc};
 use diff::{diff, diff_ignored, is_leftover_dir_with_ignored_files};
-use fs_err::{copy, create_dir, create_dir_all, read_dir, remove_dir_all, remove_file, write};
+use fs_err::{
+    copy, create_dir, create_dir_all, read_dir, remove_dir_all, remove_file, rename, write,
+};
 use portpicker::pick_unused_port;
 use rammingen::{
     cli::{Cli, Command},
@@ -19,7 +21,7 @@ use rammingen::{
     rules::Rule,
     term::{clear_status, debug, error, info},
 };
-use rammingen_protocol::ArchivePath;
+use rammingen_protocol::{util::native_to_archive_relative_path, ArchivePath};
 use rand::{seq::SliceRandom, thread_rng, Rng};
 use shuffle::{choose_path, random_content, random_name, shuffle};
 use sqlx::{query, PgPool};
@@ -113,39 +115,63 @@ async fn try_main() -> Result<()> {
     let mut snapshot_time: Option<DateTime<Utc>> = None;
     for _ in 0..1000 {
         if thread_rng().gen_bool(0.3) {
-            // upload
-            let path_for_upload = dir.join("for_upload");
-            if thread_rng().gen_bool(0.3) {
-                write(&path_for_upload, random_content())?;
-            } else {
-                create_dir(&path_for_upload)?;
-                shuffle(&path_for_upload)?;
-            }
+            // mutate through server command
             let expected = dir.join("expected");
+            if expected.exists() {
+                remove_dir_or_file(&expected)?;
+            }
             copy_dir_all(&clients[0].mount_dir, &expected)?;
-            let parent_path = choose_path(&expected, false, true, true, false)?.unwrap();
-            let path_in_expected = parent_path.join(random_name(false));
-            if path_in_expected.exists() {
-                continue;
-            }
-            if path_for_upload.is_dir() {
-                copy_dir_all(&path_for_upload, &path_in_expected)?;
+            let client1 = clients.choose(&mut thread_rng()).unwrap();
+            if thread_rng().gen_bool(0.3) {
+                // upload new path
+                let path_for_upload = dir.join("for_upload");
+                if path_for_upload.exists() {
+                    remove_dir_or_file(&path_for_upload)?;
+                }
+                if thread_rng().gen_bool(0.3) {
+                    write(&path_for_upload, random_content())?;
+                } else {
+                    create_dir(&path_for_upload)?;
+                    shuffle(&path_for_upload)?;
+                }
+                let parent_path = choose_path(&expected, false, true, true, false)?.unwrap();
+                let path_in_expected = parent_path.join(random_name(false));
+                if path_in_expected.exists() {
+                    continue;
+                }
+                if path_for_upload.is_dir() {
+                    copy_dir_all(&path_for_upload, &path_in_expected)?;
+                } else {
+                    copy(&path_for_upload, &path_in_expected)?;
+                }
+                let archive_path =
+                    archive_subpath(&archive_mount_path, &expected, &path_in_expected)?;
+                debug(format!("Checking upload ({archive_path})"));
+                client1
+                    .upload(SanitizedLocalPath::new(&path_for_upload)?, archive_path)
+                    .await?;
             } else {
-                copy(&path_for_upload, &path_in_expected)?;
+                // move path
+                let Some(path1) = choose_path(&expected, true, true, false, false)? else {
+                    continue;
+                };
+                let path2_parent = choose_path(&expected, false, true, true, false)?.unwrap();
+                let path2 = path2_parent.join(random_name(false));
+                if path2.exists() || path2.starts_with(&path1) {
+                    continue;
+                }
+                rename(&path1, &path2)?;
+                let archive_path = archive_subpath(&archive_mount_path, &expected, &path1)?;
+                let new_archive_path = archive_subpath(&archive_mount_path, &expected, &path2)?;
+                debug(format!(
+                    "Checking mv ({archive_path} -> {new_archive_path})"
+                ));
+                client1.move_path(archive_path, new_archive_path).await?;
             }
-            let archive_path = archive_subpath(&archive_mount_path, &expected, &path_in_expected)?;
-            debug(format!("Checking upload ({archive_path})"));
-            clients
-                .choose(&mut thread_rng())
-                .unwrap()
-                .upload(SanitizedLocalPath::new(&path_for_upload)?, archive_path)
-                .await?;
             for client in &clients {
                 client.sync().await?;
                 diff(&expected, &client.mount_dir)?;
             }
-            remove_dir_or_file(&expected)?;
-            remove_dir_or_file(&path_for_upload)?;
         } else {
             // edit mount
             let index = thread_rng().gen_range(0..clients.len());
@@ -159,10 +185,12 @@ async fn try_main() -> Result<()> {
                 if index2 != index {
                     debug(format!("syncing client {index2}"));
                     let before_sync_snapshot = dir.join("snapshot");
+                    if before_sync_snapshot.exists() {
+                        remove_dir_all(&before_sync_snapshot)?;
+                    }
                     copy_dir_all(&client.mount_dir, &before_sync_snapshot)?;
                     client.sync().await?;
                     diff_ignored(&client.mount_dir, &before_sync_snapshot)?;
-                    remove_dir_all(&before_sync_snapshot)?;
                 }
             }
             for client in &clients[1..] {
@@ -187,16 +215,14 @@ async fn try_main() -> Result<()> {
                     &snapshot_for_download_version_path,
                 )
                 .await?;
-                if snapshot_for_download_version_path.is_dir() {
-                    remove_dir_all(&snapshot_for_download_version_path)?;
-                } else {
-                    remove_file(&snapshot_for_download_version_path)?;
-                }
                 snapshot_time = None;
             } else {
                 info("Saving snapshot for download_version test");
                 sleep(Duration::from_millis(500)).await;
                 snapshot_time = Some(Utc::now());
+                if snapshot_for_download_version_path.exists() {
+                    remove_dir_or_file(&snapshot_for_download_version_path)?;
+                }
                 copy_dir_all(&clients[0].mount_dir, &snapshot_for_download_version_path)?;
                 sleep(Duration::from_millis(500)).await;
             }
@@ -258,6 +284,23 @@ impl ClientData {
         )
         .await
     }
+    async fn move_path(
+        &self,
+        archive_path: ArchivePath,
+        new_archive_path: ArchivePath,
+    ) -> Result<()> {
+        rammingen::run(
+            Cli {
+                config: None,
+                command: Command::Move {
+                    old_path: archive_path,
+                    new_path: new_archive_path,
+                },
+            },
+            self.config.clone(),
+        )
+        .await
+    }
 }
 
 fn archive_subpath(
@@ -269,15 +312,7 @@ fn archive_subpath(
         Ok(archive_root_path.clone())
     } else {
         let relative = path.strip_prefix(local_root_path)?;
-        let mut path = archive_root_path.clone();
-        for component in relative.components() {
-            if let path::Component::Normal(name) = component {
-                path = path.join(name.to_str().unwrap())?;
-            } else {
-                bail!("invalid path: {:?}", relative);
-            }
-        }
-        Ok(path)
+        archive_root_path.join_multiple(&native_to_archive_relative_path(relative)?)
     }
 }
 
@@ -299,6 +334,9 @@ async fn check_download(
     ));
     let client2 = clients.choose(&mut thread_rng()).unwrap();
     let destination = dir.join("tmp_download");
+    if destination.exists() {
+        remove_dir_or_file(&destination)?;
+    }
     client2
         .download(
             archive_path,
@@ -307,8 +345,6 @@ async fn check_download(
         )
         .await?;
     diff(&local_path, &destination)?;
-    remove_dir_or_file(&destination)?;
-
     Ok(())
 }
